Honor Sphinx sorting order
Using WHERE IN(3,2,1) to fetch rows from MySQL does not guarantee the sorting order is kept. In order to ensure this we have to explicitly sort by the $key field and provide the ids in the order they should be sorted in.

Since Eloquent does not support ordering by a method call we have to use orderByRaw() - which means Laravel won't handle empty matches for us like it does with the whereIn call by generating WHERE 0 = 1. Therefore I added an empty check as well which simply returns an empty collection or array. This also saves a query in case Sphinx returns no results.